backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pymongo==4.5.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
pillow==10.0.1
python-jose==3.3.0
bcrypt==4.0.1